requirements.txt
alembic==1.0.11
coverage==4.5.3
cryptography==2.7
cx-Oracle==7.2.0
flask==1.1.1
gunicorn==19.9.0
pyjwt==1.7.1
pika==1.1.0
redis==3.2.1
requests==2.22.0
selenium==3.141.0
sqlalchemy==1.3.6
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
freezegun==0.3.12
responses==0.10.6
pymssql==2.1.4
pytz==2019.3
Flask-Cors==3.0.8
newrelic==5.6.0.135
tzlocal==2.0.0
